refactor(ResultView): extract print method to ResultView class
