+++ src/main/java/racing/Car.java
@@ -3,6 +3,13 @@ package racing;
 public class Car {
     private int position;
 
+    public Car() {
+    }
+
+    public Car(int position) {
+        this.position = position;
+    }
+
     public void move() {
         position++;
     }
